Split gen3 memorycard into subfolder
